+++ Repositories/ProjectRepository.cs
@@ -65,13 +65,19 @@ namespace Repositories
         );
 
         /// <summary>
+        /// Returns project with a NON redacted User. Do not use when you do not need the user's email adress.
+        /// </summary>
+        /// <param name="id"></param>
+        /// <returns></returns>
+        Task<Project> FindAsyncNotRedacted(int id);
+
+        /// <summary>
         ///     This interface method counts the amount of projects matching the filters.
         /// </summary>
         /// <param name="highlighted"></param>
         /// <param name="categories">The categories parameter represents the categories the project needs to have</param>
-        /// <param name="userId">The user id whoms projects need to be retrieved</param>
         /// <returns>number of projects found</returns>
-        Task<int> CountAsync(bool? highlighted = null, ICollection<int> categories = null, int? userId = null);
+        Task<int> CountAsync(bool? highlighted = null, ICollection<int> categories = null);
 
         /// <summary>
         ///     This interface method searches the database for projects matching the search query and parameters.
@@ -92,7 +98,6 @@ namespace Repositories
             bool orderByAsc = true,
             bool? highlighted = null,
             ICollection<int> categories = null
-
         );
 
         /// <summary>
@@ -124,20 +129,8 @@ namespace Repositories
         ///     Get the user projects.
         /// </summary>
         /// <param name="userId">The id of the user whoms projects need to be retrieved</param>
-        /// <param name="skip">The skip parameter represents the number of projects to skip.</param>
-        /// <param name="take">The take parameter represents the number of projects to return.</param>
-        /// <param name="orderBy">The order by parameter represents the way how to order the projects.</param>
-        /// <param name="orderByAsc">The order by asc parameters represents the order direction (True: asc, False: desc)</param>
-        /// <param name="highlighted">The highlighted parameter represents the whether to filter highlighted projects.</param>
-        /// <param name="categories">The categories parameter represents the categories the project needs to have</param>
         /// <returns>A enumerable of the users projects</returns>
-        Task<IEnumerable<Project>> GetUserProjects(int userId,
-                                                   int? skip = null,
-                                                   int? take = null,
-                                                   Expression<Func<Project, object>> orderBy = null,
-                                                   bool orderByAsc = true,
-                                                   bool? highlighted = null,
-                                                   ICollection<int> categories = null);
+        Task<IEnumerable<Project>> GetUserProjects(int userId);
         Task<List<Project>> GetLikedProjectsFromSimilarUser(int userId, int similarUserId);
         void CreateProjectIndex();
         void DeleteIndex();
@@ -249,7 +242,7 @@ namespace Repositories
                 Id = p.Id,
                 ProjectIconId = p.ProjectIconId,
                 ProjectIcon = p.ProjectIcon,
-                CallToActions= p.CallToActions,
+                CallToActions = p.CallToActions,
                 Collaborators = p.Collaborators,
                 Likes = p.Likes,
                 LinkedInstitutions = p.LinkedInstitutions,
@@ -276,16 +269,9 @@ namespace Repositories
         /// </summary>
         /// <param name="highlighted">The highlighted parameter represents whether to filter highlighted projects.</param>
         /// <param name="categories">The categories parameter represents the categories the project needs to have</param>
-        /// <param name="userId">The user id whoms projects need to be retrieved</param>
         /// <returns>This method returns the amount of projects matching the filters.</returns>
-        public virtual async Task<int> CountAsync(bool? highlighted = null, ICollection<int> categories = null, int? userId = null)
+        public virtual async Task<int> CountAsync(bool? highlighted = null, ICollection<int> categories = null)
         {
-            if (userId.HasValue)
-                return await ApplyFilters(DbSet, null, null, null, true, highlighted, categories)
-                             .Where(p => p.UserId == userId)
-                             .CountAsync();
-            
-
             return await ApplyFilters(DbSet, null, null, null, true, highlighted, categories)
                        .CountAsync();
         }
@@ -482,57 +468,19 @@ namespace Repositories
         ///     Get the user projects.
         /// </summary>
         /// <param name="userId">The id of the user whoms projects need to be retrieved</param>
-        /// <param name="skip">The skip parameter represents the number of projects to skip.</param>
-        /// <param name="take">The take parameter represents the number of projects to return.</param>
-        /// <param name="orderBy">The order by parameter represents the way how to order the projects.</param>
-        /// <param name="orderByAsc">The order by asc parameters represents the order direction (True: asc, False: desc)</param>
-        /// <param name="highlighted">The highlighted parameter represents the whether to filter highlighted projects.</param>
-        /// <param name="categories">The categories parameter represents the categories the project needs to have</param>
         /// <returns>A enumerable of the users projects</returns>
-        public async Task<IEnumerable<Project>> GetUserProjects(
-            int userId,
-            int? skip = null,
-            int? take = null,
-            Expression<Func<Project, object>> orderBy = null,
-            bool orderByAsc = true,
-            bool? highlighted = null,
-            ICollection<int> categories = null)
+        public async Task<IEnumerable<Project>> GetUserProjects(int userId)
         {
-            IQueryable<Project> projects = GetDbSet<Project>()
-               .Include(p => p.Collaborators)
-               .Include(p => p.ProjectIcon)
-               .Include(p => p.Images)
-               .Include(p => p.Categories)
-               .ThenInclude(c => c.Category)
-               .Where(p => p.UserId == userId);
+            IEnumerable<Project> projects = await GetDbSet<Project>()
+                   .Include(p => p.Collaborators)
+                   .Include(p => p.ProjectIcon)
+                   .Include(p => p.Images)
+                   .Include(p => p.Categories)
+                   .ThenInclude(c => c.Category)
+                   .Where(p => p.UserId == userId)
+                   .ToListAsync();
 
-            projects = ApplyFilters(projects, skip, take, orderBy, orderByAsc, highlighted, categories);
-
-            List<Project> projectResults = await projects.Select(p => new Project
-            {
-                   UserId = p.UserId,
-                   User = p.User,
-                   Id = p.Id,
-                   ProjectIconId = p.ProjectIconId,
-                   ProjectIcon = p.ProjectIcon,
-                   CallToActions = p.CallToActions,
-                   Collaborators = p.Collaborators,
-                   Likes = p.Likes,
-                   LinkedInstitutions = p.LinkedInstitutions,
-                   Categories = p.Categories.Select(c => new ProjectCategory()
-                       {
-                           Category = c.Category,
-                           Id = c.Id
-                       }).ToList(),
-                   Created = p.Created,
-                   InstitutePrivate = p.InstitutePrivate,
-                   Name = p.Name,
-                   ShortDescription = p.ShortDescription,
-                   Updated = p.Updated,
-                   Uri = p.Uri
-            })
-                .ToListAsync();
-            return projectResults;
+            return projects;
         }
 
         /// <summary>
@@ -615,7 +563,6 @@ namespace Repositories
         /// <param name="orderByAsc">The order by asc parameters represents the order direction (True: asc, False: desc)</param>
         /// <param name="highlighted">The highlighted parameter represents the whether to filter highlighted projects.</param>
         /// <param name="categories"></param>
-        /// <param name="userId">The user id whoms projects need to be retrieved</param>
         /// <returns>
         ///     This method returns a IQueryable Projects collection based on the given filters.
         /// </returns>
@@ -626,8 +573,7 @@ namespace Repositories
             Expression<Func<Project, object>> orderBy,
             bool orderByAsc,
             bool? highlighted,
-            ICollection<int> categories,
-            int? userId = null
+            ICollection<int> categories
         )
         {
             if(highlighted.HasValue)
@@ -659,11 +605,6 @@ namespace Repositories
                 {
                     queryable = queryable.OrderByDescending(orderBy);
                 }
-            }
-
-            if(userId.HasValue)
-            {
-                queryable = queryable.Where(p => p.UserId == userId);
             }
 
             if(skip.HasValue && skip.Value > 0) queryable = queryable.Skip(skip.Value);
@@ -838,10 +779,9 @@ namespace Repositories
             return GetDbSet<Project>().AnyAsync(i => i.Id == id);
         }
 
-        public Task<int> CountAsync(bool? highlighted = null, ICollection<int> categories = null)
+        public Task<Project> FindAsyncNotRedacted(int id)
         {
-            throw new NotImplementedException();
+            return GetDbSet<Project>().Where(p => p.Id == id).Include(u => u.User).FirstOrDefaultAsync();
         }
     }
-
 }