+++ Services/Services/ProjectService.cs
@@ -50,20 +50,25 @@ namespace Services.Services
         Task<Project> FindWithUserCollaboratorsAndInstitutionsAsync(int id);
 
         /// <summary>
+        /// Returns project with a NON redacted User. Do not use when you do not need the user's email adress.
+        /// </summary>
+        /// <param name="id"></param>
+        /// <returns></returns>
+        Task<Project> FindAsyncNotRedacted(int id);
+
+        /// <summary>
         ///     Get the number of projects
         /// </summary>
         /// <param name="projectFilterParams">The parameters to filter, sort and paginate the projects</param>
-        /// <param name="userId">The user id whoms projects need to be retrieved</param>
         /// <returns>The number of projects</returns>
-        Task<int> ProjectsCount(ProjectFilterParams projectFilterParams, int? userId = null);
+        Task<int> ProjectsCount(ProjectFilterParams projectFilterParams);
 
         /// <summary>
         ///     Get the total number of pages for the results
         /// </summary>
         /// <param name="projectFilterParams">The parameters to filter, sort and paginate the projects</param>
-        /// <param name="userId">The user id whoms projects need to be retrieved</param>
         /// <returns>The total number of pages for the results</returns>
-        Task<int> GetProjectsTotalPages(ProjectFilterParams projectFilterParams, int? userId = null);
+        Task<int> GetProjectsTotalPages(ProjectFilterParams projectFilterParams);
 
         Task<bool> ProjectExistsAsync(int id);
 
@@ -71,9 +76,8 @@ namespace Services.Services
         ///     Get the users projects
         /// </summary>
         /// <param name="userId">The user id whoms projects need to be retrieved</param>
-        /// <param name="projectFilterParams">The parameters to filter, sort and paginate the projects</param>
         /// <returns>The total number of pages for the results</returns>
-        Task<IEnumerable<Project>> GetUserProjects(int userId, ProjectFilterParams projectFilterParams);
+        Task<IEnumerable<Project>> GetUserProjects(int userId);
         /// <summary>
         ///     Registers all records of the current database to the message broker to be added to ElasticSearch.
         /// </summary>
@@ -184,14 +188,9 @@ namespace Services.Services
         ///     Get the number of projects
         /// </summary>
         /// <param name="projectFilterParams">The parameters to filter, sort and paginate the projects</param>
-        /// <param name="userId">The user id whoms projects need to be retrieved</param>
         /// <returns>The number of projects</returns>
-        public virtual async Task<int> ProjectsCount(ProjectFilterParams projectFilterParams, int? userId = null)
+        public virtual async Task<int> ProjectsCount(ProjectFilterParams projectFilterParams)
         {
-            if(userId.HasValue)
-            {
-                return await Repository.CountAsync(projectFilterParams.Highlighted, projectFilterParams.Categories, userId);
-            }
             return await Repository.CountAsync(projectFilterParams.Highlighted, projectFilterParams.Categories);
         }
 
@@ -200,12 +199,12 @@ namespace Services.Services
         /// </summary>
         /// <param name="projectFilterParams">The parameters to filter, sort and paginate the projects</param>
         /// <returns>The total number of pages for the results</returns>
-        public virtual async Task<int> GetProjectsTotalPages(ProjectFilterParams projectFilterParams, int? userId = null)
+        public virtual async Task<int> GetProjectsTotalPages(ProjectFilterParams projectFilterParams)
         {
             if(projectFilterParams.AmountOnPage == null ||
                projectFilterParams.AmountOnPage <= 0)
                 projectFilterParams.AmountOnPage = 20;
-            int count = await ProjectsCount(projectFilterParams, userId);
+            int count = await ProjectsCount(projectFilterParams);
             return (int) Math.Ceiling(count / (decimal) projectFilterParams.AmountOnPage);
         }
 
@@ -260,48 +259,18 @@ namespace Services.Services
         }
 
 
-        public Task<IEnumerable<Project>> GetUserProjects(int userId, ProjectFilterParams projectFilterParams)
+        public Task<IEnumerable<Project>> GetUserProjects(int userId)
         {
-            if(!projectFilterParams.AmountOnPage.HasValue ||
-               projectFilterParams.AmountOnPage <= 0)
-                projectFilterParams.AmountOnPage = 20;
-
-            int? skip = null;
-            int? take = projectFilterParams.AmountOnPage;
-            if(projectFilterParams.Page.HasValue && projectFilterParams.Page.Value > 1)
-            {
-                skip = projectFilterParams.AmountOnPage * (projectFilterParams.Page - 1);
-                take = projectFilterParams.AmountOnPage;
-            }
-
-            Expression<Func<Project, object>> orderBy;
-            switch(projectFilterParams.SortBy)
-            {
-                case "name":
-                    orderBy = project => project.Name;
-                    break;
-                case "created":
-                    orderBy = project => project.Created;
-                    break;
-                case "likes":
-                    orderBy = project => project.Likes.Count;
-                    break;
-                default:
-                    orderBy = project => project.Updated;
-                    break;
-            }
-
-            bool orderByDirection = projectFilterParams.SortDirection == "asc";
-            return Repository.GetUserProjects(userId, skip,
-                                              take,
-                                              orderBy,
-                                              orderByDirection,
-                                              projectFilterParams.Highlighted,
-                                              projectFilterParams.Categories);
+            return Repository.GetUserProjects(userId);
         }
         public async Task<List<Project>> FindProjectsWhereTitleStartsWithQuery(string query)
         {
             return await Repository.FindProjectsWhereTitleStartsWithQuery(query);
+        }
+
+        public async Task<Project> FindAsyncNotRedacted(int id)
+        {
+            return await Repository.FindAsyncNotRedacted(id);
         }
     }